+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added subscript setter for FREArray i.e. myFreArray[0] = myFREObject
 - Added iterator for FREArray i.e. for freObject in myFreArray { }
 - Added subscript for setProp i.e. myFreObject["propName"] = myFREObject
+- Improve UIImage Extension
 
 ### 2.4.0
 - Upgraded to Xcode 9.3